new project in this repo
Had to actually start a new project as my laptop had some issues.
Found it easier to redo the project than to try and solve the issues as it would have been mor time consuming.

Added functionality for member registration (add and update).
To add donation and tax receipts to complete the flow.